check-ref-format: allow --branch outside repository again
"git check-ref-format --branch $name" feature was originally
introduced (and was advertised) as a way for scripts to take any
end-user supplied string (like "master", "@{-1}" etc.) and see if
that is usable when Git expects to see a branch name, and also
obtain the concrete branch name that the at-mark magic expands to.

We made the command fail outside a repository, where the latter half
of the feature does not make sense because "the branch we were on
before the last checkout" is only meanfull in a repository,

However, the former half of the feature is still useful even outside
a repository, because "git check-ref-format --branch $name" can be
stricter than "git check-ref-format refs/heads/$name", and Porcelain
scripts may want to check a given name against the stricter rule.

Allow the command to start outside a repository, but avoid running
the at-mark magic expansion when doing so.

Signed-off-by: Junio C Hamano <gitster@pobox.com>