Restore old behaviour of push --recurse-submodules
This reverts commit 06bf4ad1db92c32af38e16d9b7f928edbd647780.

The reverted commit allowed the remote and refspec arguments to `git
push` to be propagated to submodules. One effect of this was to prevent
pushing submodules whose checked-out branch is different from the
supermodule. The old behaviour was essentially to `git push` the
submodules, relying on a branch (with properly set-up tracking branch)
being checked out in each submodule.

The fact that I've reverted the commit is not a judgement on its
quality. I am not contributing this to the mainline, and I don't think
it should ever be merged. However, my team relies on the old behaviour
of recurse-submodules=on-demand and we don't pass a refspec for that
operation.

3 relevant failing tests have been left in.